+++ editor/src/components/canvas/grid-panels-container.tsx
@@ -17,6 +17,7 @@ import { LeftPaneComponent } from '../navigator/left-pane'
 import { CodeEditorPane, ResizableRightPane } from './design-panel-root'
 import { useGridPanelDragInfo, useGridPanelDropArea } from './grid-panels-dnd'
 import type { Menu, Pane, PanelData } from './grid-panels-state-2'
+import { GridPanel } from './grid-panel'
 
 export type PanelName = Menu | Pane
 
@@ -42,8 +43,8 @@ type StoredLayout = Array<Array<StoredPanel>>
 const NumberOfPanels = 4
 const IndexOfCanvas = 2
 
-const VerticalGapHalf = 6
-const HorizontalGapHalf = 6
+export const GridPanelVerticalGapHalf = 6
+export const GridPanelHorizontalGapHalf = 6
 
 const ExtraHorizontalDropTargetPadding = 15
 
@@ -282,12 +283,12 @@ export const GridPanelsContainer = React.memo(() => {
         gridTemplateColumns: `[col] ${columnWidths[0]}px [col] ${columnWidths[1]}px [canvas] 1fr [col] ${columnWidths[2]}px [col] ${columnWidths[3]}px [end]`,
         gridTemplateRows: 'repeat(12, 1fr)',
         gridAutoFlow: 'dense',
-        columnGap: HorizontalGapHalf * 2,
-        rowGap: VerticalGapHalf * 2,
-        paddingTop: VerticalGapHalf,
-        paddingBottom: VerticalGapHalf,
-        paddingLeft: HorizontalGapHalf + 2,
-        paddingRight: HorizontalGapHalf + 2,
+        columnGap: GridPanelHorizontalGapHalf * 2,
+        rowGap: GridPanelVerticalGapHalf * 2,
+        paddingTop: GridPanelVerticalGapHalf,
+        paddingBottom: GridPanelVerticalGapHalf,
+        paddingLeft: GridPanelHorizontalGapHalf + 2,
+        paddingRight: GridPanelHorizontalGapHalf + 2,
       }}
     >
       <GridPanel
@@ -380,9 +381,9 @@ const ColumnDragTargets = React.memo(
             gridRowEnd: -1,
             gridColumn: `col ${columnIndex > -1 ? columnIndex + 1 : columnIndex} / span 1`,
             display: isDragActive && canDropBefore ? 'block' : 'none',
-            width: 2 * ExtraHorizontalDropTargetPadding + 2 * HorizontalGapHalf,
+            width: 2 * ExtraHorizontalDropTargetPadding + 2 * GridPanelHorizontalGapHalf,
             height: '100%',
-            left: -(ExtraHorizontalDropTargetPadding + 2 * HorizontalGapHalf),
+            left: -(ExtraHorizontalDropTargetPadding + 2 * GridPanelHorizontalGapHalf),
           }}
         >
           <div
@@ -405,9 +406,9 @@ const ColumnDragTargets = React.memo(
             gridRowEnd: -1,
             gridColumn: `col ${columnIndex > -1 ? columnIndex + 1 : columnIndex} / span 1`,
             display: isDragActive && canDropAfter ? 'block' : 'none',
-            width: 2 * ExtraHorizontalDropTargetPadding + 2 * HorizontalGapHalf,
+            width: 2 * ExtraHorizontalDropTargetPadding + 2 * GridPanelHorizontalGapHalf,
             height: '100%',
-            right: -(ExtraHorizontalDropTargetPadding + 2 * HorizontalGapHalf),
+            right: -(ExtraHorizontalDropTargetPadding + 2 * GridPanelHorizontalGapHalf),
           }}
         >
           <div
@@ -425,171 +426,3 @@ const ColumnDragTargets = React.memo(
     )
   },
 )
-
-interface GridPanelProps {
-  onDrop: (itemToMove: StoredPanel, newPosition: LayoutUpdate) => void
-  canDrop: (itemToMove: StoredPanel, newPosition: LayoutUpdate) => boolean
-  pane: PanelData
-}
-
-export const GridPanel = React.memo<GridPanelProps>((props) => {
-  const { onDrop, canDrop } = props
-  const { panel, index, span, order } = props.pane
-
-  const { isDragActive, draggedPanel } = useGridPanelDragInfo()
-
-  const dropAboveElement: LayoutUpdate = React.useMemo(
-    () => ({
-      type: 'before-index',
-      indexInColumn: order,
-      columnIndex: index,
-    }),
-    [order, index],
-  )
-
-  const dropBelowElement: LayoutUpdate = React.useMemo(
-    () => ({
-      type: 'after-index',
-      indexInColumn: order,
-      columnIndex: index,
-    }),
-    [order, index],
-  )
-
-  const canDropAbove = draggedPanel != null && canDrop(draggedPanel, dropAboveElement)
-  const canDropBelow = draggedPanel != null && canDrop(draggedPanel, dropBelowElement)
-
-  const { drop: dropBefore, isOver: isOverBefore } = useGridPanelDropArea(
-    index,
-    order,
-    React.useCallback(
-      (itemToMove: StoredPanel) => onDrop(itemToMove, dropAboveElement),
-      [onDrop, dropAboveElement],
-    ),
-  )
-  const { drop: dropAfter, isOver: isOverAfter } = useGridPanelDropArea(
-    index,
-    order,
-    React.useCallback(
-      (itemToMove: StoredPanel) => {
-        onDrop(itemToMove, dropBelowElement)
-      },
-      [onDrop, dropBelowElement],
-    ),
-  )
-
-  const draggablePanelComponent = (() => {
-    switch (panel.name) {
-      case 'code-editor':
-        return (
-          <CodeEditorPane
-            panelData={props.pane.panel}
-            small={false}
-            width={0}
-            height={0}
-            onResize={NO_OP}
-            setIsResizing={NO_OP}
-            resizableConfig={{
-              enable: {
-                right: true,
-              },
-            }}
-          />
-        )
-      case 'inspector':
-        return (
-          <ResizableRightPane
-            panelData={props.pane.panel}
-            width={0}
-            height={0}
-            onResize={NO_OP}
-            setIsResizing={NO_OP}
-            resizableConfig={{
-              snap: {
-                x: [UtopiaTheme.layout.inspectorSmallWidth, UtopiaTheme.layout.inspectorLargeWidth],
-              },
-              enable: {
-                left: true,
-              },
-            }}
-          />
-        )
-      case 'navigator':
-        return (
-          <LeftPaneComponent
-            panelData={props.pane.panel}
-            width={0}
-            height={0}
-            onResize={NO_OP}
-            setIsResizing={NO_OP}
-            resizableConfig={{
-              minWidth: LeftPanelMinWidth,
-              enable: {
-                right: true,
-              },
-            }}
-          />
-        )
-      default:
-        return null
-    }
-  })()
-
-  return (
-    <div
-      style={{
-        pointerEvents: 'initial',
-        gridColumn: `col ${index > -1 ? index + 1 : index}`,
-        gridRow: `span ${span}`,
-        order: order,
-        display: 'flex',
-        flexDirection: 'column',
-        contain: 'layout',
-      }}
-    >
-      {draggablePanelComponent}
-      <div
-        ref={dropBefore}
-        style={{
-          display: isDragActive && canDropAbove ? 'block' : 'none',
-          position: 'absolute',
-          width: '100%',
-          height: `calc(50% + ${VerticalGapHalf}px)`,
-          top: -VerticalGapHalf,
-        }}
-      >
-        <div
-          style={{
-            display: isOverBefore ? 'block' : 'none',
-            position: 'absolute',
-            top: -1,
-            height: 2,
-            width: '100%',
-            backgroundColor: colorTheme.primary.value,
-          }}
-        />
-      </div>
-      <div
-        ref={dropAfter}
-        style={{
-          display: isDragActive && canDropBelow ? 'block' : 'none',
-          position: 'absolute',
-          width: '100%',
-          height: `calc(50% + ${VerticalGapHalf}px)`,
-          bottom: -VerticalGapHalf,
-        }}
-      >
-        <div
-          style={{
-            display: isOverAfter ? 'block' : 'none',
-            position: 'absolute',
-            bottom: -1,
-            height: 2,
-            width: '100%',
-            backgroundColor: colorTheme.primary.value,
-          }}
-        />
-      </div>
-    </div>
-  )
-})
